requirements.txt
arxiv
requests
schedule
openai
python-dotenv
jinja2
beautifulsoup4
lxml
reportlab
